+++ src/sentry/api/endpoints/user_notification_details.py
@@ -7,9 +7,9 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
 from sentry.api.serializers import Serializer, serialize
 from sentry.notifications.types import UserOptionsSettingsKey
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.models.user_option import UserOption
 
 USER_OPTION_SETTINGS = {
