+++ src/sentry/api/endpoints/user_notification_email.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.models.user_option import UserOption
 from sentry.users.models.useremail import UserEmail
