+++ src/sentry/api/endpoints/user_notification_settings_options_detail.py
@@ -6,8 +6,8 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
 from sentry.models.notificationsettingoption import NotificationSettingOption
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.models.user import User
