+++ src/sentry/users/api/endpoints/user_emails_confirm.py
@@ -6,9 +6,9 @@ from rest_framework.response import Response
 
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
 from sentry.api.validators import AllowedEmailField
 from sentry.types.ratelimit import RateLimit, RateLimitCategory
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.models.user import User
 from sentry.users.models.useremail import UserEmail
