+++ src/sentry/users/api/endpoints/user_identity_config.py
@@ -8,9 +8,9 @@ from rest_framework.response import Response
 
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserAndStaffPermission, UserEndpoint
 from sentry.api.serializers import serialize
 from sentry.models.authidentity import AuthIdentity
+from sentry.users.api.bases.user import UserAndStaffPermission, UserEndpoint
 from sentry.users.api.serializers.user_identity_config import (
     Status,
     UserIdentityConfig,
