+++ src/sentry/users/api/endpoints/user_ips.py
@@ -3,10 +3,10 @@ from rest_framework.response import Response
 
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
 from sentry.api.decorators import sudo_required
 from sentry.api.paginator import DateTimePaginator
 from sentry.api.serializers import serialize
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.api.serializers.userip import UserIPSerializer
 from sentry.users.models.user import User
 from sentry.users.models.userip import UserIP
