+++ src/sentry/users/api/endpoints/user_permissions_config.py
@@ -5,8 +5,8 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint
 from sentry.api.permissions import SuperuserOrStaffFeatureFlaggedPermission
+from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.models.user import User
