+++ src/sentry/users/api/endpoints/user_regions.py
@@ -8,12 +8,12 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
-from sentry.api.bases.user import UserEndpoint, UserPermission
 from sentry.auth.superuser import is_active_superuser
 from sentry.auth.system import is_system_auth
 from sentry.models.organizationmapping import OrganizationMapping
 from sentry.models.organizationmembermapping import OrganizationMemberMapping
 from sentry.types.region import get_region_by_name
+from sentry.users.api.bases.user import UserEndpoint, UserPermission
 from sentry.users.models.user import User
 from sentry.users.services.user import RpcUser
